created admin panel

+++ app/Http/Controllers/CategoriesController.php
@@ -98,6 +98,10 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts->count() > 0){
+            session()->flash('error', 'Category cannot be deleted because it has some posts.');
+            return redirect()->back();
+        }
         $category->delete();
         session()->flash('success', 'Category Delete Successfully');
         return redirect(route('categories.index'));
